pyproject.toml
[tool.poetry]
name = "r_place"
version = "0.1.0"
description = "Create an r/place timelapse of your area"
authors = ["zevaryx <zevaryx@gmail.com>"]
license = "Expat License"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
